pyproject.toml
[tool.poetry]
name = "diplom-bulletin-board"
version = "0.1.0"
description = ""
authors = ["alexandra kashina <kashin159@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
redis = "^5.2.1"
black = "^24.10.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
flake8 = "^7.1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
ipython = "^8.31.0"
coverage = "^7.6.10"
isort = "^5.13.2"
drf-yasg = "^1.21.8"
stripe = "^11.4.1"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
